+++ src/resources/index.ts
@@ -19,3 +19,4 @@ export { L4OrderBookResource } from './l4-orderbook';
 export { L2OrderBookResource } from './l2-orderbook';
 export { L3OrderBookResource } from './l3-orderbook';
 export { SpotPairsResource, SpotTwapResource } from './spot';
+export { Hip3BreadthResource } from './hip3-breadth';